move isDesktop and isMilliseconds to Helpers.ts
